// harvestrackr-next/src/types/database.ts
export type PaymentMethod =
  | "CASH"
  | "CHECK"
  | "CREDIT_CARD"
  | "DEBIT_CARD"
  | "BANK_TRANSFER"
  | "ONLINE"
  | "OTHER";